requirements.txt
Flask>=3.0.0
Flask-Migrate>=4.0.0
Flask-SQLAlchemy>=3.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0